+++ Highlights26.md
@@ -1,0 +1,5 @@
+# What's new in Anorm 2.6
+
+This page highlights the new features of Anorm 2.6. If you want learn about the changes you need to make to migrate to Anorm 2.6, check out the [[Migration Guide|Migration25]].
+
+- Iteratees module
